Pass bundles
Separates existing passes into two separate types:
  - passes - standard passes that transform IR
  - bundles - passes that have no custom logic and only run pass
              pipelines

Additionally, the existing bundles are moved to individual source files
to make repo navigation easier. Plus some minor cleanup.

The goal is to improve pipeline modularization and make it easier to
navigate available tools within TPP-MLIR.
