+++ include/TPP/PassUtils.h
@@ -18,13 +18,26 @@
 namespace mlir {
 namespace tpp {
 
-// Helper base class for passes that call and manage combination of other
+// Helper base class for bundle passes that call and manage combination of other
 // existing passes.
-template <typename OpT> class UtilityPassBase {
+template <typename OpT = void> class PassBundle {
 public:
-  UtilityPassBase()
+  PassBundle()
       : pm(OpT::getOperationName(), mlir::OpPassManager::Nesting::Implicit){};
-  virtual ~UtilityPassBase() = default;
+  virtual ~PassBundle() = default;
+
+protected:
+  OpPassManager pm;
+
+  // Create the pass processing pipeline.
+  virtual void constructPipeline() = 0;
+};
+
+// Pass bundle specialization without anchor operation type.
+template <> class PassBundle<void> {
+public:
+  PassBundle() : pm(mlir::OpPassManager::Nesting::Implicit){};
+  virtual ~PassBundle() = default;
 
 protected:
   OpPassManager pm;
