+++ lib/TPP/Bundles/CMakeLists.txt
@@ -1,17 +1,13 @@
-add_subdirectory(Bundles)
-add_subdirectory(Dialect)
-add_subdirectory(Conversion)
-add_subdirectory(IR)
-add_subdirectory(GPU)
-add_subdirectory(Runner)
-add_subdirectory(Transforms)
-
 get_property(mlir_dialect_libs GLOBAL PROPERTY MLIR_DIALECT_LIBS)
 get_property(conversion_libs GLOBAL PROPERTY MLIR_CONVERSION_LIBS)
 
-add_mlir_library(TPPPipeline
-  DefaultPipeline.cpp
-  DefaultTppPasses.cpp
+add_mlir_library(TPPPassBundles
+  Cleanup.cpp
+  LinalgLowering.cpp
+  LocalDialectsLowering.cpp
+  LowLevelParallelization.cpp
+  PostProcessing.cpp
+  TppMapping.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${PROJECT_SOURCE_DIR}/include/TPP
@@ -24,6 +20,4 @@ add_mlir_library(TPPPipeline
     MLIRIR
     ${mlir_dialect_libs}
     ${conversion_libs}
-    TPPGPU
-    TPPPassBundles
   )
