+++ lib/TPP/Bundles/PostProcessing.cpp
@@ -1,0 +1,58 @@
+//===- PostProcessing.cpp ----------------------------------------*- C++-*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#include "TPP/Bundles.h"
+
+#include "mlir/Dialect/Bufferization/IR/Bufferization.h"
+#include "mlir/Dialect/Bufferization/Transforms/Passes.h"
+#include "mlir/Dialect/Func/IR/FuncOps.h"
+#include "mlir/Dialect/MemRef/IR/MemRef.h"
+#include "mlir/Dialect/SCF/IR/SCF.h"
+#include "mlir/IR/BuiltinOps.h"
+#include "mlir/Pass/Pass.h"
+#include "mlir/Pass/PassManager.h"
+
+#include "TPP/PassUtils.h"
+
+using namespace mlir;
+using namespace mlir::tpp;
+
+namespace mlir {
+namespace tpp {
+#define GEN_PASS_DEF_POSTPROCESSING
+#include "TPP/Bundles.h.inc"
+} // namespace tpp
+} // namespace mlir
+
+// Apply various postprocessing passes such as LICM, parallel loop fusion,
+// buffer deallocation, general cleanup etc.
+struct Postprocessing : public tpp::impl::PostprocessingBase<Postprocessing>,
+                        PassBundle<func::FuncOp> {
+  using PostprocessingBase::PostprocessingBase;
+
+  void runOnOperation() override {
+    auto module = getOperation();
+
+    // Initialize the pipeline if needed.
+    // Otherwise, just run the cached one.
+    if (pm.empty())
+      constructPipeline();
+
+    if (failed(runPipeline(pm, module)))
+      return signalPassFailure();
+  }
+
+private:
+  void constructPipeline() override {
+    // Postprocess buffers.
+    pm.addPass(bufferization::createBufferHoistingPass());
+
+    // Run general cleanup to normalize IR.
+    pm.addPass(createCleanup());
+  }
+};
