+++ tools/mlir-gen/mlir-gen.cpp
@@ -1,4 +1,10 @@
-//===- mlir-gen MLIR Generator --------------------------------------------===//
+//===- mlir-gen MLIR Generator -----------------------------------*- C++-*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
 //
 // Main entry-point to the MLIR generator. Creates an MLIR model with input,
 // output and multiple hidden layers, different activation functions, etc.
